+++ test/config.ts
@@ -1,5 +1,8 @@
+import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
+import { StepTestConfig } from '@jupiterone/integration-sdk-testing';
 import * as dotenv from 'dotenv';
 import * as path from 'path';
+import { invocationConfig } from '../src';
 import { IntegrationConfig } from '../src/config';
 
 if (process.env.LOAD_ENV) {
@@ -11,3 +14,11 @@ if (process.env.LOAD_ENV) {
 export const config: IntegrationConfig = {
   apiToken: process.env.API_TOKEN || 'apiToken',
 };
+
+export function buildStepTestConfigForStep(stepId: string): StepTestConfig {
+  return {
+    stepId,
+    instanceConfig: config,
+    invocationConfig: invocationConfig as IntegrationInvocationConfig,
+  };
+}